Function intel_tee_quote_verification_rs::sgx_qv_set_path

pub fn sgx_qv_set_path(
    path_type: sgx_qv_path_type_t,
    path: &str
) -> quote3_error_t

Set the full path of QVE and QPL library.
The function takes the enum and the corresponding full path.

Param

path_type
The type of binary being passed in.

path
It should be a valid full path.

Return

SGX_QL_SUCCESS
Successfully set the full path.

SGX_QL_ERROR_INVALID_PARAMETER
Path is not a valid full path or the path is too long.